+++ moon/vehicles/rover.py
@@ -60,7 +60,8 @@ import math
 from datetime import timedelta
 
 from osgar.lib.mathex import normalizeAnglePIPI
-from osgar.node import Node
+
+from moon.moonnode import MoonNode
 
 
 WHEEL_RADIUS = 0.275  # meters
@@ -71,12 +72,10 @@ WHEEL_NAMES = ['fl', 'fr', 'bl', 'br']
 
 CRAB_ROLL_ANGLE = 0.78
 
-class Rover(Node):
+class Rover(MoonNode):
     def __init__(self, config, bus):
         super().__init__(config, bus)
         bus.register('cmd', 'pose2d')
-
-        self.sim_time = None
 
         # general driving parameters
         # radius: radius of circle to drive around, "inf" if going straight; 0 if turning round in place
@@ -101,9 +100,6 @@ class Rover(Node):
         self.yaw = 0.0
         self.yaw_offset = None
         self.in_driving_recovery = False
-
-    def on_sim_clock(self, data):
-        self.sim_time = timedelta(seconds=data[0], microseconds=data[1] // 1000)
 
     def on_driving_recovery(self, data):
         self.in_driving_recovery = data
@@ -281,15 +277,6 @@ class Rover(Node):
 
         cmd = b'cmd_rover %f %f %f %f %f %f %f %f' % tuple(steering + effort)
         self.bus.publish('cmd', cmd)
-
-    def update(self):
-        channel = super().update()
-        handler = getattr(self, "on_" + channel, None)
-        if handler is not None:
-            handler(getattr(self, channel))
-
-        return channel
-
 
     def draw(self):
         # for debugging
